Use systemd-run --user to launch
This work! Wow!

I think, maybe, now that I reflect more, that env vars set inside a piped command
(like the while read -r line) don't actually get to the shell, that they're being
run in a subshell? So the eval export VAR=value looks good but doesn't have impact.
This discrepancy can be observed looking at the `env` debugging statement removed
in this comit, which showed that the attempts to re-import the env weren't having
the expected impact.

I'd like to figure out a solution that doesn't rely on systemd-run but I'm also
pretty damned happy having systemd-run there too. The downside though is that
the systemd-run isn't going to redirect stdin/stdout &c & log! It will return
right away. And that's not the idea at all.